Fix `io.test_for_file` failing for editable installs
The `io.test_for_file` function would fail to find files that are
actually present in the `pdb2pqr/dat` folder if the package was
installed in editable mode. The reason is that in editable mode the
files are not actually copied to a path that is in the `PYTHONPATH` but
the source tree is simply symlinked. Since the code was only checking
base directories that are the current working directory or part of
`sys.path`, files were never found.

There the absolute path of `pdb2pqr/dat` is determined by simply getting
the filepath of the `pdb2pqr.io` module and then getting the `dat`
folder through relative path operations. This should work for all manner
of installing.